Move Attachment utils to separate files
