+++ apps/melding-form/src/app/(general)/bijlage/_utils/startUpload.ts
@@ -1,23 +1,6 @@
 import type { Dispatch, SetStateAction } from 'react'
 
-export const safeJSONParse = (jsonString?: string) => {
-  if (!jsonString) return undefined
-
-  try {
-    return JSON.parse(jsonString)
-  } catch {
-    return undefined
-  }
-}
-
-export const VALIDATION_ERROR_MESSAGES_TRANSLATION_KEYS: Record<string, string> = {
-  'Allowed content size exceeded': 'validation-errors.file-too-large',
-  'Attachment not allowed': 'validation-errors.invalid-file-type',
-  'Media type of data does not match provided media type': 'validation-errors.invalid-file-extension',
-}
-
-export const getValidationErrorMessageTranslationKey = (error?: string): string =>
-  (error && VALIDATION_ERROR_MESSAGES_TRANSLATION_KEYS[error]) || 'validation-errors.failed-upload'
+import { safeJSONParse } from '~/app/_utils/safeJSONParse'
 
 export type FileUpload = {
   errorMessage?: string
@@ -33,6 +16,15 @@ export type PendingFileUpload = Omit<FileUpload, 'status'> & {
   status: 'pending'
   xhr: XMLHttpRequest
 }
+
+export const VALIDATION_ERROR_MESSAGES_TRANSLATION_KEYS: Record<string, string> = {
+  'Allowed content size exceeded': 'validation-errors.file-too-large',
+  'Attachment not allowed': 'validation-errors.invalid-file-type',
+  'Media type of data does not match provided media type': 'validation-errors.invalid-file-extension',
+}
+
+const getValidationErrorMessageTranslationKey = (error?: string): string =>
+  (error && VALIDATION_ERROR_MESSAGES_TRANSLATION_KEYS[error]) || 'validation-errors.failed-upload'
 
 // We're using XMLHttpRequest instead of fetch here,
 // because fetch does not allow you to track the upload progress.
@@ -52,17 +44,19 @@ export const startUpload = (
   }
 
   xhr.onload = () => {
+    type Response = { detail?: string; id?: number }
+
+    const parsed = safeJSONParse<Response, undefined>(xhr.response, undefined)
+    const isOk = xhr.status >= 200 && xhr.status < 300
+
     setFileUploads((prev) =>
       prev.map((upload): FileUpload | PendingFileUpload =>
         upload.id === fileUpload.id
           ? {
               ...upload,
-              errorMessage:
-                xhr.status !== 200
-                  ? getValidationErrorMessageTranslationKey(safeJSONParse(xhr.response)?.detail)
-                  : undefined,
-              serverId: safeJSONParse(xhr.response)?.id,
-              status: xhr.status !== 200 ? 'error' : 'success',
+              errorMessage: isOk ? undefined : getValidationErrorMessageTranslationKey(parsed?.detail),
+              serverId: parsed?.id,
+              status: isOk ? 'success' : 'error',
             }
           : upload,
       ),
